Merge from upstream

+++ tools/wsdlto/core/src/main/java/org/apache/cxf/tools/wsdlto/WSDLToJavaContainer.java
@@ -32,6 +32,7 @@ import java.io.OutputStream;
 import java.io.Writer;
 import java.net.URI;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -370,29 +371,32 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
     private void createClientJar(File tmpDirectory, JarOutputStream jarout) {
         try {
             URI parentFile = new File((String)context.get(ToolConstants.CFG_CLASSDIR)).toURI();
-            for (File file : tmpDirectory.listFiles()) {
-                URI relativePath = parentFile.relativize(file.toURI());
-                String name = relativePath.toString();
-                if (file.isDirectory()) {
-                    if (!StringUtils.isEmpty(name)) {
-                        if (!name.endsWith("/")) {
-                            name += "/";
+            File[] files = tmpDirectory.listFiles();
+            if (files != null) {
+                for (File file : files) {
+                    URI relativePath = parentFile.relativize(file.toURI());
+                    String name = relativePath.toString();
+                    if (file.isDirectory()) {
+                        if (!StringUtils.isEmpty(name)) {
+                            if (!name.endsWith("/")) {
+                                name += "/";
+                            }
+                            JarEntry entry = new JarEntry(name);
+                            entry.setTime(file.lastModified());
+                            jarout.putNextEntry(entry);
+                            jarout.closeEntry();
                         }
-                        JarEntry entry = new JarEntry(name);
-                        entry.setTime(file.lastModified());
-                        jarout.putNextEntry(entry);
-                        jarout.closeEntry();
+                        createClientJar(file, jarout);
+                        continue;
                     }
-                    createClientJar(file, jarout);
-                    continue;
+                    JarEntry entry = new JarEntry(name);
+                    entry.setTime(file.lastModified());
+                    jarout.putNextEntry(entry);
+                    InputStream input = new BufferedInputStream(new FileInputStream(file));
+                    IOUtils.copy(input, jarout);
+                    input.close();
+                    jarout.closeEntry();
                 }
-                JarEntry entry = new JarEntry(name);
-                entry.setTime(file.lastModified());
-                jarout.putNextEntry(entry);
-                InputStream input = new BufferedInputStream(new FileInputStream(file));
-                IOUtils.copy(input, jarout);
-                input.close();
-                jarout.closeEntry();
             }
         } catch (Exception e) {
             Message msg = new Message("FAILED_ADD_JARENTRY", LOG);
@@ -553,11 +557,22 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
         if (!env.containsKey(ToolConstants.CFG_WSDLLOCATION)) {
             //make sure the "raw" form is used for the wsdlLocation
             //instead of the absolute URI that normalize may return
+            boolean assumeFileURI = false;
             try {
                 URI uri = new URI(wsdl);
+
+                String uriScheme = uri.getScheme();
+                if (uriScheme == null) {
+                    assumeFileURI = true;
+                }
+
                 wsdl = uri.toString();
             } catch (Exception e) {
                 //not a URL, assume file
+                assumeFileURI = true;
+            }
+
+            if (assumeFileURI) {
                 if (wsdl.indexOf(":") != -1 && !wsdl.startsWith("/")) {
                     wsdl = "file:/" + wsdl;
                 } else {
@@ -570,6 +585,7 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
                     //ignore... 
                 }
             }
+
             wsdl = wsdl.replace("\\", "/");
 
             env.put(ToolConstants.CFG_WSDLLOCATION, wsdl);
@@ -606,9 +622,8 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
     protected void setLibraryReferences(ToolContext env) {
         Properties props = loadProperties(getResourceAsStream("wsdltojavalib.properties"));
         if (props != null) {
-            for (Iterator<?> keys = props.keySet().iterator(); keys.hasNext();) {
-                String key = (String)keys.next();
-                env.put(key, props.get(key));
+            for (Map.Entry<Object, Object> entry : props.entrySet()) {
+                env.put((String)entry.getKey(), entry.getValue());
             }
         }
         env.put(ToolConstants.CFG_ANT_PROP, props);
@@ -665,7 +680,7 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
             file.delete();
             File tmpFile = file.getParentFile();
             while (tmpFile != null && !tmpFile.getCanonicalPath().equalsIgnoreCase(outPutDir)) {
-                if (tmpFile.isDirectory() && tmpFile.list().length == 0) {
+                if (tmpFile.isDirectory() && tmpFile.list() != null && tmpFile.list().length == 0) {
                     tmpFile.delete();
                 }
                 tmpFile = tmpFile.getParentFile();
@@ -679,7 +694,8 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
                 classFile.delete();
                 File tmpClzFile = classFile.getParentFile();
                 while (tmpClzFile != null && !tmpClzFile.getCanonicalPath().equalsIgnoreCase(outPutDir)) {
-                    if (tmpClzFile.isDirectory() && tmpClzFile.list().length == 0) {
+                    if (tmpClzFile.isDirectory() && tmpClzFile.list() != null 
+                        && tmpClzFile.list().length == 0) {
                         tmpClzFile.delete();
                     }
                     tmpClzFile = tmpClzFile.getParentFile();
@@ -699,10 +715,7 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
             || context.optionSet(ToolConstants.CFG_GEN_FAULT)) {
             return true;
         }
-        if (context.optionSet(ToolConstants.CFG_NO_TYPES)) {
-            return true;
-        }
-        return false;
+        return context.optionSet(ToolConstants.CFG_NO_TYPES);
     }
 
     public void generateTypes() throws ToolException {
@@ -808,7 +821,7 @@ public class WSDLToJavaContainer extends AbstractCXFToolContainer {
                     Element el = imp.getSchemaDocument().getDocumentElement();
                     updateImports(el, sourceMap);
                     os = new FileWriterUtil(impfile.getParent(), context.get(OutputStreamCreator.class))
-                        .getWriter(impfile, "UTF-8");
+                        .getWriter(impfile, StandardCharsets.UTF_8.name());
                     StaxUtils.writeTo(el, os, 2);
                     os.close();
                 }
